# tests/contracts/openapi-v1.json
{
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonScalar"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
      "Body_photosUpload": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_photosUpload",
        "type": "object"
      },
      "Category": {
        "description": "Supported complaint categories; OTHER requires human review.",
        "enum": [
          "pothole",
          "blocked_drain",
          "flooding",
          "rubbish",
          "street_light",
          "other"
        ],
        "title": "Category",
        "type": "string"
      },
      "ClusteringStatus": {
        "enum": [
          "confirmed",
          "isolated",
          "conflict"
        ],
        "title": "ClusteringStatus",
        "type": "string"
      },
      "ComplaintCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Category"
              },
              {
                "type": "null"
              }
            ]
          },
          "latitude": {
            "description": "Latitude in decimal degrees.",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "description": "Longitude in decimal degrees.",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "photo_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Identifier returned by the photo upload endpoint.",
            "title": "Photo Id"
          },
          "photo_path": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Legacy free-text photo reference; server-managed uploads use photo_id.",
            "title": "Photo Path"
          },
          "reported_at": {
            "description": "Timezone-aware report time, normalized to UTC.",
            "format": "date-time",
            "title": "Reported At",
            "type": "string"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "latitude",
          "longitude",
          "reported_at"
        ],
        "title": "ComplaintCreateRequest",
        "type": "object"
      },
      "ComplaintResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "complaint_id": {
            "format": "uuid",
            "title": "Complaint Id",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "photo_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Path"
          },
          "reported_at": {
            "format": "date-time",
            "title": "Reported At",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "complaint_id",
          "text",
          "category",
          "latitude",
          "longitude",
          "reported_at"
        ],
        "title": "ComplaintResponse",
        "type": "object"
      },
      "ComplaintSubmissionResponse": {
        "additionalProperties": false,
        "properties": {
          "complaint": {
            "$ref": "#/components/schemas/ComplaintResponse"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "incident_transition": {
            "$ref": "#/components/schemas/IncidentTransitionResponse"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummaryResponse"
            },
            "title": "Incidents",
            "type": "array"
          },
          "priorities": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PriorityResponse"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Priorities",
            "type": "array"
          },
          "relationship_decisions": {
            "items": {
              "$ref": "#/components/schemas/RelationshipEdgeResponse"
            },
            "title": "Relationship Decisions",
            "type": "array"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          }
        },
        "required": [
          "complaint",
          "created",
          "replayed",
          "relationship_decisions",
          "incident_transition",
          "incidents",
          "priorities"
        ],
        "title": "ComplaintSubmissionResponse",
        "type": "object"
      },
...
      "GeoPointResponse": {
        "additionalProperties": false,
        "properties": {
          "latitude": {
            "description": "Latitude in decimal degrees.",
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "description": "Longitude in decimal degrees.",
            "title": "Longitude",
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "GeoPointResponse",
        "type": "object"
      },
...
      "IncidentSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "category_summary": {
            "items": {
              "$ref": "#/components/schemas/Category"
            },
            "title": "Category Summary",
            "type": "array"
          },
          "centroid": {
            "$ref": "#/components/schemas/GeoPointResponse"
          },
          "confirmed_report_count": {
            "title": "Confirmed Report Count",
            "type": "integer"
          },
          "conflict_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Conflict Reasons",
            "type": "array"
          },
          "earliest_reported_at": {
            "format": "date-time",
            "title": "Earliest Reported At",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "latest_reported_at": {
            "format": "date-time",
            "title": "Latest Reported At",
            "type": "string"
          },
          "pending_candidate_count": {
            "title": "Pending Candidate Count",
            "type": "integer"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PriorityResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "radius_metres": {
            "title": "Radius Metres",
            "type": "number"
          },
          "status": {
            "$ref": "#/components/schemas/ClusteringStatus"
          }
        },
        "required": [
          "incident_id",
          "status",
          "category_summary",
          "priority",
          "confirmed_report_count",
          "pending_candidate_count",
          "centroid",
          "radius_metres",
          "earliest_reported_at",
          "latest_reported_at",
          "conflict_reasons"
        ],
        "title": "IncidentSummaryResponse",
        "type": "object"
      },
      "IncidentTransitionResponse": {
        "additionalProperties": false,
        "properties": {
          "current_incident_snapshot_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Current Incident Snapshot Ids",
            "type": "array"
          },
          "previous_incident_snapshot_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Previous Incident Snapshot Ids",
            "type": "array"
          }
        },
        "required": [
          "previous_incident_snapshot_ids",
          "current_incident_snapshot_ids"
        ],
        "title": "IncidentTransitionResponse",
        "type": "object"
      },
      "JsonScalar": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "PhotoUploadResponse": {
        "additionalProperties": false,
        "properties": {
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "photo_id": {
            "format": "uuid",
            "title": "Photo Id",
            "type": "string"
          }
        },
        "required": [
          "photo_id",
          "media_type",
          "byte_size"
        ],
        "title": "PhotoUploadResponse",
        "type": "object"
      },
...
      "PriorityResponse": {
        "additionalProperties": false,
        "properties": {
          "level": {
            "$ref": "#/components/schemas/PriorityLevel"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "level",
          "reasons",
          "policy_version"
        ],
        "title": "PriorityResponse",
        "type": "object"
      },
...
      "RelationshipEdgeResponse": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/MatchState"
          },
          "decision_source": {
            "$ref": "#/components/schemas/RelationshipDecisionSource"
          },
          "left_id": {
            "format": "uuid",
            "title": "Left Id",
            "type": "string"
          },
          "matcher_recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchState"
              },
              {
                "type": "null"
              }
            ]
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "right_id": {
            "format": "uuid",
            "title": "Right Id",
            "type": "string"
          }
        },
        "required": [
          "left_id",
          "right_id",
          "decision",
          "reasons",
          "decision_source",
          "matcher_recommendation"
        ],
        "title": "RelationshipEdgeResponse",
        "type": "object"
      },
...
      "SeedResetResponse": {
        "additionalProperties": false,
        "properties": {
          "complaint_count": {
            "title": "Complaint Count",
            "type": "integer"
          },
          "incident_count": {
            "title": "Incident Count",
            "type": "integer"
          },
          "priority_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Priority Counts",
            "type": "object"
          },
          "review_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Review Counts",
            "type": "object"
          },
          "seed_checksum": {
            "title": "Seed Checksum",
            "type": "string"
          },
          "seed_version": {
            "title": "Seed Version",
            "type": "string"
          }
        },
        "required": [
          "seed_version",
          "seed_checksum",
          "complaint_count",
          "incident_count",
          "review_counts",
          "priority_counts"
        ],
        "title": "SeedResetResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CivicPulse-lite API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/reset": {
      "post": {
        "description": "Admin operation: restore the server-configured demo seed; disabled by default.",
        "operationId": "adminReset",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedResetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Administrative reset is disabled."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Seed configuration is unavailable."
          }
        },
        "summary": "Reset Seed",
        "tags": [
          "mutations"
        ]
      }
    },
    "/api/v1/complaints": {
      "post": {
        "description": "Submit a complaint through the application mutation service.",
        "operationId": "complaintsCreate",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintSubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Idempotency conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Submit Complaint",
        "tags": [
          "mutations"
        ]
      }
    },
...
    "/api/v1/photos": {
      "post": {
        "description": "Upload JPEG or PNG photo evidence. The media type is detected from the file content; the server assigns the stored name.",
        "operationId": "photosUpload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_photosUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Photo exceeds the size cap."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unsupported photo content."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Photo persistence failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Photo storage is unavailable."
          }
        },
        "summary": "Upload Photo",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/photos/{photo_id}": {
      "get": {
        "description": "Serve stored photo evidence with its recorded media type.",
        "operationId": "photosGet",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Photo Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Stored JPEG or PNG photo evidence."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Photo not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Photo identifier validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Photo retrieval failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Photo storage is unavailable."
          }
        },
        "summary": "Get Photo",
        "tags": [
          "photos"
        ]
      }
    },
...
  }
}